Make CI jobs executor parameterized
We have two different executors. One basic and one which includes a full
features setup for the end-to-end tests. While the whole installation
and build process happens in the basic executor, the tests run in the
later one. This does not work out correctly and it require to re-install
the dependencies in the new image to make them fully available (note
that their files are already shipped via attached workspace). Therefore
it is the new approach to install and build everything from the start in
the end-to-end image for the workflow including the according tests.
To solve that, the jobs to install and build have become parameterized,
allowing the caller to specify the executor. That remains per default
the basic executor.