Use and document new e2e DockerHub repository
There has been a new DockerHub repository created, following the new
naming scheme for the lightclient-e2e-environment. Therefore the
documentation has become updated to link to this image. Else the script
to run the end-to-end tests got updated to pull this image automatically
rather than requiring the user to have it built before. Anyways the user
is still able to use a local built.