Improve e2e script for MacOS and being more rebust
MacOS is again interpreting a standard Unix command differently. This
also showed that the script might can fail, causing the started Docker
container to eventually not get stopped and removed. Therefore we the
script now exists for every error a child could throw and runs
a trap exit function to shut down Docker in any case.