dApp: require mainnet to be explicitly enabled
Before we required it to be explicitly disabled. While doing that for
productive builds at the moment, this might fail if something goes wrong
with the environment build. Since it is more harmful to have mainnet
accidentally enabled than not having it available, it has been agreed
that it makes more sense to require the enabling explicitly. All
environments do that already for both cases. So it is just a safety net.