circleci: add parameter for dApp build mode
The if-else logic in the job to build the dApp is not well placed. The
job should not decide when to do what. This especially has been shown
when we discovered that the publish workflow has build the dApp in
development mode not ghpages. It makes more sense to parametrize the job
and let the workflow control what should be build how.
Notice that for this commit, the publish workflow has been fixed, but
the staging version is in development mode now. This needs to be fixed.
But since this is related to a more general problem of bad structured
workflows, this will be part of another commit.