circleci: improve workflow filters and publish job
The management of adopting the behavior of the workflows is quite
a mess. So there is a "hidden" workflow that behaves differently on the
master branch. That one is publishing a dApp build to the staging
sub-directory. To fix that and also make sure the staging build is done
with the correct mode applied, the following improvements have been
done:

Workflows are straight forward and don't change their behavior depending
on the branch/tag. Thereby each different behavior needs to be reflected
by a separate workflow. Unfortunately it is not yet possible to apply
filter properties to workflows at top level (yet). Rather they have to
be applied to the jobs in the flow. But since all jobs depend in
a tree/chain on the checkout job, it is easy to apply the filter here.

In result of the above paragraph, the mass of filters applied to the
jobs in the publish workflow could have been removed and applied to the
checkout job only.

A new workflow for the master branch, publishing a build in ghpages mode
to the staging subdirectory has been added. The default workflow no more
publish anything conditionally. This was also necessary to define the
mode in which the build should be done, which differs between the
default case and staging.

The publish workflow has been renamed to be more complained with the
situation to have multiple publish workflows now.

The deploy.sh script was dependent on the $DEPLOY_BASE variable,
injected via CircleCI contexts. This is super hard to see and
understand, requiring the developer to access/read the CircleCI
organisation settings etc. Therefore this has been changed to just
interpret the first call argument as the deploy base. This allows to be
independent of a hard to get context and also offers the opportunity to
expose a job parameter to the workflows. Thereby the publish staging
workflow can easily tell the publish job to put the dApp build into the
'staging' subdirectory. On the other side the publish release workflow
acts on the root directory of the ghpages webserver.