sdk: add force exit parameter to CI e2e tests
The end-to-end tests running for the SDK in the CI do succeed, but are
not exiting correctly. Therefore CircleCI reports a failed job in the
workflow, even the actual tests are positive. This appears to be
related to the current issue of the SDK with WebRTC during shutdown.
This is caused by a bug in the used library that needs to be fixed.
Until then we tell Jest to force exit the SDK after the tests have run
through. Thereby the overall job will be successful again.
This change could eventually hide other issues (although they should be
super unlikely). So we should remove this flag as soon as the WebRTC
library got fixed.