dApp: Cypress test scenarios (#2232)
* Test for accepting disclaimer

A very first Cypress test which acccepts the disclaimer and asserts that the user is redirected to the home route.

* Introducing fixtures

This is a first attempt to make sense about how to split the tests into reusable chunks. Here accessing the disclaimer route has been split into a separate function while the assertions happens in the test for the disclaimer.

* Renames fixtures file and function name

* Tests from launching dApp to connecting

The conditional in the connectToDApp function is there to make it possible to reuse that function even when a Subkey already has been generated.

* Hook to delete indexedDBs before a test run

Cypress supports hooks that runs before and after each test and the documentation provides examples on how to add them. Therefore I added the code for deleting indexedDBs in a before hook. The documentation also suggests to use the hooks inside the scope of a describe block.

* Tests to open channel

* Remaining tests refactoring

* Kebab-case custom event

Latest es-lint update requires custom events to be in kebab-case, this one was not and it is fixed now.

* Ignore uncaught exceptions in Cypress

Cypress does not recognize things like @ts-ignore that we ourselves have put in the code and fails the test suite because of that.

* Runs all refactored test functions

* Change to shortname for flag

Mac only recognizes the shortname for directory.

* Deleting boilerplate comments

* TEMPORALLY ENABLE NIGHLTIES FOR PR TESTING

* Naming anonymous function

This makes it easier to understand what the function for ingoring uncought exceptions actually does.

Co-authored-by: Thore Weilbier <thore@weilbier.net>