dapp: fix correct match of build modes and paths
The dApp gets build for different modes. These modes define the
environment(-variables) of the build. But they are also used for many
other conditions during the build process or the code itself. So is also
the public path property determined by it. The public path is required
to tell the packager of a single page application (SPA) at which
"sub-directory" the assets will be served later. On older versions this
variable was also called the base URL. So already at build time it must
be known where the page will become served.

The issue was that this public path was not set correctly. This happened
after the recent fix for the correct usage of modes for the CI different
workflows. This previous bug was kinda hiding the not so well defined
public path determination. As result the staging build was using the
productive environment with a public path set to '/', although it
becomes deployed at '/staging/'. Because the SPA was packed wrongly, it
tried to access assets at root level that do not exist there, but only
in the sub-path.

This whole issue was supported by not using a unique naming of the same
things, a mix of terminology by framework and project plus some bad
written code. So here come the fixes:

The Vue CLI differs per default in 'production', 'development' and
'test' modes. More modes can be always added or existing overwritten
(but then require a manual adoption for all commands). The project has
et al. introduced the 'ghpages' mode. In fact it is equal to what Vue
CLI understands as 'production'. This 'ghpages' mode was loading the
environment correctly, but the determination of the public path was
broken, while being based on the default mode names. Furthermore do the
release (how the CI calls it) and staging builds share the same mode
('ghpages'), even the build needs to differ between both.
The public path determination was kinda shady written. To solve that,
there was a new optional variable introduced that defines the public
path and can be set for each mode individually. Per default (as Vue CLI
does define it) the build is meant for production (public path is '/').
The 'ghpages' mode got now removed and replaced with 'production' and
'staging'. They are clones of 'ghpages', but differ the definition of
the public path variable. This also resulted into a renaming of the
`deploy_release` and `deploy_staging` CI workflows to
`deploy_production` and `deploy_staging`.

Furthermore there was another mixed terminology usage. While the "base
URL" is called public path in the Vue CLI build, it was called deploy
base for the custom GitHub pages deploy script. Although they must refer
to the same value for build and deployment. Therefore this has become
equalized to public path for both steps.

To make the CI worklows even more clear to read, all parameters that
configure the jobs have become required. This means they have no more
a default value and the worklows need to specify their value explicitly
(else CI fails for an invalid configuration).

In relation to the former change, the public path (former deploy base)
parameter for the deploy script is now accepting leading slashes. Before
we used an empty string for the root directory of the production
deployment (former release/ghpages). To make it a little bit more
obvious to read the workflow defines this parameter as '/' (or
'/staging/' for staging) now. The deploy script has been simply adopted
to handle this case properly.