serve_pr: fix for new parametrized job names
The job to build the dApp at CircleCI has become parametrized. This
causes the job name to change by getting an appendix with a dash and
a number. Thereby the serve script was not able to find the job and
downloading the built artifact. This was solved by searching for the job
with a regex following the above described pattern.