sdk: fetch presence explicitly on presence events
This works around wrong presence events being sent by the server for
some users on higher-load scenarios. It's a little bit wasteful to make
a new explicit request for it, but it seems this endpoint is proxied
through the federation and will return the actual presence despite what
the local server sees.