dApp: add both ETH balances to account overview
The AccountContent view is displaying a single ETH balance. It gives no
information from which account this balance is. Moreover does this
balance depend on if the Raiden account has a balance or not.
To improve this view and do not confuse the user there are two
improvements. First should the balance not switch dependent on if there
is a Raiden account balance or not, but rather when a Raiden account is
in use or not. But actually it is more informative for the user to just
show both balances. In case no Raiden account is used, the main account
balance is sufficient.