dApp: switch UDC screen ETH balance for account
The UDC screen in the account router shows an ETH balance during the
withdrawal procedure. The relevant balance depends on if a Raiden
account is used. So far it always shows the main account balance. This
can be confusing, when the Raiden account has no balance, the
transaction to withdrawal fails, but the UI displays a positive (and
high enough) ETH balance. Therefore this balance must always relate to
the correct account that is relevant for this transaction.