ci: fix not executed jobs for production workflow
During the last release procedure we discovered an issue with our
updated CircleCI configuration. The problem was that the production
workflow was successfully triggered, but only the first job got
executed.
It has turned out that tags behave differently in CircleCI. While for
branches we use the approach that apply a filter to the first job in the
workflow only and organize all other jobs in a dependent sequence, this
does not work for tags. Tags trigger per default nothing. And also if
the checkout job of the production workflow was executed cause it was
explicitly filtered for release tags, the dependent jobs like install
etc. where not executed, since they had not such filter. Each job that
should get executed on a tag must get the same filter repeat.
This is an ugly design, but we can't change it. The only optimization we
can do is to recover the former YAML anchor to make the same filter
getting re-used. This is also the suggestion of the official
documentation by CircleCI.