ci: merge checkout and install job to optimize
There is never the case that we want to run the checkout job without
running install directory afterwards. Therefore it makes sense to merge
both jobs into one. The advantages are that we keep the configuration
and workflow descriptions shorter and a performance boost. Each jobs
needs to get prepared by CircleCI. This costs time. Since the overall
runtime of the default worklow is already quite long, such improvements
without any drawbacks are really welcome.