ci: use pipeline parameters instead of anchors
For values that are shared between many sections in the configuration we
use YAML anchors to declare them once and then inject everywhere. For
the basic key value scenario, CircleCI provides this features as well.
Such is called pipeline parameters. The advantage of these parameters is
that they can be overwritten when triggering jobs or workflows via the
API. This is especially useful for testing new setups (e.g. adopt the
docker image version of an executor). Also it can be argued that this is
the preferable syntax. It also supports type checks.
Though there are still some anchors for more complex YAML expressions
that are not just key value pairs.