ci: make use of run working directory property
The run descriptions of a step inside a job has the option to define
a working directory. So far we don't use it and rather do a manual
change directory command in the command string. This makes the command
property hard to read and less focused. Also where the paths not
relative to the working directory of the job, which makes it partially
useless.