ci: define report directory for executors
The directory where to store directories is defined as environment
variable on each job which requires it. Since this is duplicated code
this is better set on the executor definition, shared between all jobs.
Furthermore this showed that the most properties of the both executors
are the same and can be shared via a YAML anchor.
The actual path of the report directory is also used in the command to
upload the test results. Therefore the path became a new pipeline
variable.

This commit also removes a bunch of quotation marks from the
configuration. They are usually not necessary for simple key-value-pairs
in YAML and just cause not wished highlighting. Adding quotes in YAML is
usually reserved for special cases where necessary or uncertain about
the interpreter.