ci: fix report/artifact upload for sdk tests
Apparently it got unnoticed that the upload of the test results for the
SDK tests did not work in CircleCI (if it ever did).
The reason therefore was that the SDK has a configuration for the Junit
reporter where to place its report output. Apparently this takes
precedence over the environment variable defined by the CircleCI
executor. Since the report target was awkwardly set to the coverage
output directory for the SDK, this got simply removed. Thereby it will
point per default to the current working directory executing the tests.
Within the continuous environment, the variable is taken now and the
upload of the report and artifact works again.

The SDK had already a pretty exhaustive configuration for its test
reporters in Jest. Therefore it was useless to list them again as CLI
arguments within the CI job. As result it made sense to also introduce
some configuration for Jest in the dApp. This allows to get rid of the
CLI parameters here as well and it provides a nice test report per
default.