dApp: Prepare state backup notifications with link
These preparations allows us to add a notification for reminding users to back up their state. The notification includes a link that redirects the user to the backup state menu. Since the i18n are defined in the raiden-service.ts and passed as strings to the component my workaround for providing a link was to add a getter with a switch case which maps to a navigation mixin.