sdk: export ESLint ignore rules to own file
All ESLint configurations have become exported into separate files. This
was introduced because we now use Yarn workspaces and share a lot of
configuration in the root. The only left snippet were the ignore rules
within the SDK workspace. Since this is an exception to the rest, this
should get exported to an `.eslintignore` file. Thereby no linting or
formatting rules are left within any of the package description files.