dapp: fix linting issues
Due to the recent uniformation of linting rules between all workspaces
in the repository, a mass of errors have been reported for the dApp,
which has used more loose rules in the past.
This commit fixes the errors and warning within the source directory of
the dApp. A lot of changes where necessary. It is not worth to explain
all of them here. After all the functionality hasn't changed and it was
mainly focused on the type-system anyways.
The remaining linting issues are remaining in the tests directory only.