dApp: State backup notification (#2298)
* dApp: Prepare state backup notifications with link

These preparations allows us to add a notification for reminding users to back up their state. The notification includes a link that redirects the user to the backup state menu. Since the i18n are defined in the raiden-service.ts and passed as strings to the component my workaround for providing a link was to add a getter with a switch case which maps to a navigation mixin.

* dApp: Backup state reminder mutation and state

We extend the part of the state which persists the disclaimer with an additional value for remembering a date as number. This date will be used to determine when to trigger new notification reminder about backing up the state.

* dApp: Backup state reminder mutation and state

* dApp: Notify user about backing up state

A new notification for backing up state is pushed when a user connects if the last time the user was notified was more than 24h ago.

* dApp: Organize notification icons in sub dir

* dApp: Replaces switch statement with router.push

* dApp: Better code organization for notification

* dApp: Updated variable names

* dApp: Move notification data to function

* dApp: Tests for backup state notify dates

* Trying to pass CI by updating spy function in test

* Tests for notification links

* Fixes build issues

* Changelog entry