ci: use npx to run license checker
The tool to check licenses within the CI pipeline is not part of the
development dependencies for any workspace. Rather it gets installed
only for the job to check the licenses. The current approach is pretty
cumbersome. A way more simple solution is to simply make use of the npx
command that allows to execute binaries from the npm registry on the
fly, while caching version between the runs. This simplifies the related
job and barely makes the used command more complex.