dapp: fix reactivity of notification description
The description of a notification gets handled by its own component. It
bring the feature to detect Ethereum addresses within the text and treat
it special.
The issue with this component was that it broke the reactivity. This
means that if the notification its description changed while it was
visible, the displayed text did not. This was caused by the component to
give the description some special treatment when it gets mounted. But
afterwards it never updates anymore and shows only the initial property
value. This got simply fixed to just observe the property and rerun the
parsing of the text again.