dapp: improve blockchain related notifications
There are notifications which are related to transactions on the
blockchain. They can either fail right away (which means there was never
a transaction) or succeed. Both cases are identified by a different kind
of action emitted by the SDK. The notification send for the successful
case includes additional information according to the transaction. This
also allows the notification card to show a counter of remaining block
until the associated transaction gets confirmed. When the counter
reaches zero it simply assumes it was successful. But this must not be
the case.
The SDK emits each action multiple times for each new block. Initially
the confirmation state is undefined. The last emitted action will then
either show it as confirmed (true) or there was a reorg on the
blockchain (false). This does not get used at the moment at all.
Therefore the following improvements have been done. The notification
card independently shows the counter if possible/needed. The
RaidenService who updates the notification will alternate the
description if the transaction is still about to get confirmed or was
confirmed. Finally if there is a success action, but the confirmation
state becomes false, the notification gets replaced with the error
version of it (as for the failure action). Here the description gets
filled with the reason for the failure (in this case the reorg).
The notification card itself should not figure out itself if it is
confirmed or not.

These changes have been made to the channel open related actions only.
This is so, since the whole transaction relation stuff was only
introduced here. We need to extend this to other notifications as well.