dapp: remove async modifiers for address getters
The RaidenService class forwards the access to the main and Raiden
address form the SDK. For some unknown reason the service module made
these functions asynchronous. Even the SDK source isn't asynchronous.
They are plain getters in Typescript.
This design was causing the need to make some lifecycle functions of Vue
components asynchronous. This is especially an issue for the `mount`
hook that causes problems during the tests. Therefore this should get
removed now and work plain synchronous.

It is the future goal to make these both functions pure getters in the
dApp as it is already in the SDK. But this requires way more refactoring
due to the special mocking in Jest for the unit tests. Since this commit
just got branched out from a current feature development, it is hard to
ensure that this refactoring is correctly working and therefore gets
postponed.