dapp: split store for new udc module
This is a commit in preparation for planned future features. Thereby
there will come a bunch of new data to the store that is related to the
user deposit contract. Since the root store is already cluttered and the
code hard to maintain, it has been decided that it makes more sense to
introduce a new module for the store as we did for the notifications.
While this commit does not add any new state, the user deposit token
address and the token itself got extracted to the new module. This
brings the change that tokens registered at the registry contract and
the token for the user deposit contract are handled separately. This
reduces a lot of confusion that took place in the past. Note that if the
token used for the user deposit contract has a registered token network,
this will still work as before. Just the handling is different now.