dapp: fix components and test for new udc store
This commit got split to make the actual feature of the new user deposit
store module in the previous commit better to read.

This commit includes all the adoptions in the components, modules and
tests to be compliant with this new module. Usually this is just an
adoption of the Vuex state mappers in the components.
While adopting the tests, it ha been discovered that many of them still
use the store itself. According to the principle of unit tests this
should not be done. Following the agreement of the development team,
such tests suites should be adopted over time when they get touched.
Thereby this commit became quite large to improve those test
suites.