dapp: extend udc store for planned withdrawal
This extend the Vuex store module for user deposit related data to
include the currently planned withdraw. Such get added and update via
the events of the SDK. If a withdraw plan successful took place it gets
added to the store in pending mode. From there it can either get updated
to become confirmed or it gets removed due to a reorg of the blockchain.
If the plan becomes confirmed and the user withdraws the tokens, it gets
removed form the stores state as well.
There can always be only one withdraw plan according to the contract
implementation. Therefore we don't care about old data and always just
overwrite.
The SDK were in need to expose the relevant events to make the dApp
working.