dapp: add component for planned udc withdrawals
The state of a possibly currently planned withdrawal form the user
deposit contract should be shown at the accounts UDC screen. This commit
implements the designed component with all its features plus tests.
This commit does not care about the correct injection to the UDC screen.
This is target of another issue that will rework the structure of this
screen. Therefore this might look a little bit off for the moment.