dApp: Creating InfoOverlay component
Component for the info overlay. It simply gets the correct data from the routes meta field and displays it in the components template. At the moment the placeholder image is used instead of the image defined in the route meta. This will be replaced in a future PR when the designer has provided the final designs.