dapp: remove navigation from UDC withdrawal info
We have encountered that it makes no sense to provide the user
a navigation link from a ready to claim planned user deposit withdrawal.
The claim happens automatically. Plus there will be tokens to see on the
withdrawal screen at this point. And as soon as the automatic claim
happens, the planned UDC information blocks disappears and shows that
there is no planned one. Therefore this link was never practically
useful at any time.
The removal of the link allowed for some small simplifications of the
plugin.