dapp: make notification card size more flexible
Also we attempt to make most notifications take the same size, some will
have longer text combined with a link. After all it simply looks
horrible when the text does not fit into the box. So we should rather
have a larger box first and eventually adjust the content later, than
having a "broken" UI in-between. Therefore I adjusted the fixed height
to a minimum height. This guarantees for the best case that all cards
have the same height, but in case it is needed they can grow to fit
their content.