dapp: resolve all errors and warning in unit tests
There were a couple of warnings an errors when running the unit tests.
But they were to weak to let the tests fail. Anyways they are errors and
clutter the tests output, hiding actual new errors. Else the developer
gets desensitized for red error output.

One of the main changes was the new dependency which brings a mock
implementation for HTML canvas in Jest. Since we can use this simply
globally in the Jest setup, this is an easy fix that applies everywhere.

The result is a wall of green output for the dApp their unit tests.
Unfortunately does the Vue cli plugin for the unit tests with Jest still
has some warning about incompatible versions. This gets printed at the
very beginning of the execution and is only a hint. And since it is not
resolvable on our side, we must keep it like that for the moment.