dapp: fix notification card for undefined icon
The notification payload property of an icon is optional. But the
notification card component treats it as required. Since the type checks
in templates are not good enough, this was undiscovered and let to
unexpected failures were notifications were not shown due to an
undefined property.
The workaround adds a fallback icon that always gets used when the
payload does not specify such. This also moves the type check to the
components class code where type checks are available.