dapp: add initial pre-caching to service worker
The requirement is that the dApp is fully functional without the serving
web-server, once the dApp was loaded by the user. To fulfill this
requirement, each new service worker (so always after a new build dApp)
will pre-cache all assets of the server when become activated. Therefore
we use the manifest as generated by Webpack about all assets. This
happens in parallel while the user is already using the dApp. At the
moment the this are about 8.1 MB of data to fetch and cache. Depending
on the network connection that should be done pretty fast.