dapp: extend raiden shutdown to navigate back home
When the RaidenService gets stopped the current state gets reset. After
that the application is not operable anymore until the user has
re-connected again. Therefore he needs to go to the home/connect screen.
This does not happen automatically at the moment. Therefore this just
has been added.
So far the service was basically never shutdown except for when the user
closes the application. There we could never see that this was not
working as it was also not required. But if the service stops during
runtime, this is necessary to do.

This is in preparation to the upcoming update features. Before an update
the RaidenService must be safely stopped.