dapp: use workbox directory instead of vue plugin
The vue-cli-pwa plugin is causing issues. First it is still stuck on an
ancient version of workbox (> two major versions). Second it does
some things in background we don't want to, don't understand, but also
shouldn't.
Furthermore does it "magically" inject the workbox packages to the
service worker during compilation. This causes us linting issues. But
more critical it is a security issues (detected by Firefox) as it is
loaded externally instead of getting bundled with the build.
After all it is just too easy to use the workbox packages ourself
manually and have the full control over it.