dapp: register service worker via vue plugin
The package `register-service-worker` is causing some issues. It does
not much by itself and should just provide a thin convenient wrapper
around standard functionality. But it is inconvenient since it renames
the events of the W3 service worker specification. Furthermore it
behaves slightly different on localhost which is a nightmare for
testing.
After all we need barely any feature of the package. It is super easy to
register the service worker manually.

The approach to create a new Vue plugin happens with the future in mind.
The service worker runs in a separate thread of the applications mains
thread with strictly separated data access. Therefore the service worker
will need an assistant in the main thread it can communicate with. And
components will need to get access to this (e.g. when a user wants to
trigger an update). Thereby it makes sense to use a plugin here.

Note that the update event caught by the update snackbar component is
broken now. This happens on purpose and will become replaced later
completely.