dapp: prevent automatic update by service worker
The precache feature of workbox we use fetches the most recent version
during installation (and then later activates it) of the service
worker. This means each time a new service worker gets installed by the
browser, this will update the locally served sources.
This means the "update" of the service worker and the application are
bound together at the moment. This is critical, because we don't have
full control of when the service worker gets updated. In fact service
workers are inteded to get automatically updated by the webbrowser.
Therfore do browsers update to the most recent version each 24 hours,
disregarding any caching time set by the server. This meant to let users
never stuck on an eventual broken service worker.
For us the requirements differ. The user must actively acknowledge that
he wants to update the application. Furthermore does the update function
of service workers behave unhappiely for us. The function will always
fetch the file from the server and then does a byte to byte comparison
of the files. As it is now, each time the application gets build, this
result into a new "version" of the worker script, no matter if the
application got updated or not.
After all it is necessary to decouple the update of the service worker
and the actual updates of the application itself (means the files in the
cache that get served). But we still want to use the workbox features of
precaching since doing it ourself is way too much work and error prone.
But we can make use of its smaller components instad of using the
one-shot function that does all. Therefore we manually create the
precache controller and router instances. Then we manually call the
lifecycle hook functions if necessary. Necessary means when there is
actually an update required. For the moment we only handle the
first-time-loaded state. But we need to have the future update procedure
in mind. Due to the bounding to the lifecycle of the service worker we
are faced with the problem that the "old" service worker must inform the
next service worker that he should run the hooks to install the most
recent version. The only data structure the worker thread has access to
is the cache API. The resulting simple solution is to clear the cache.
This means when a new service worker registers an empty cache, it will
fill this cache with the most recent version. This solves the starting
issue when the page gets loaded the first time, but also allows the
"old" service worker to cause the next service worker to update.

The service worker script has been moved into its own source directory.
There is more about to come and it makes sense to group them together.
The sub-directory is inteded to group the pure JavaScript files
together. It is for the moment not possible to compile a single source
directory with the TypeScript webworker library *and* the dom library.
Therfore the service worker remains JS until this upstream issue gets
solved (issue open).

The service worker is using so called extendable events. Event handler
can hold worker thread within a certain event state until a promise
resolves. This for exmaple makes sure that the service worker does not
get activated before it has completed the install phase (which could
take a while since it downloads all sources). But functions which run
within these extended events do have a different context. Since it is
necessary to persist a temporally state between the lifecycle hooks, all
functions must be called with the correct context set. This could get
solved by using arrow functions. But since we are using a utilitiy
module to keep the service worker module clean to read, this won't work
for those functions. Therefore will need to bind the context anyways.