dapp: add generation of version variable and file
This the preparation to allow the application to find out if there is an
update available or not. Therefore the current package its version gets
injected into the application as an environment variable. Furthermore
a version file gets generated during the packaging and placed next to
the other static assets. This allows any client to fetch this version
file of the server to know which version it is serving. Thereby the
application can later check which version is available on the server and
compare it with itself.
Note that the version file will not be cached by the service worker,
which would prevent the client to see new versions.

It was necessary to disable ESLint for importing webpack. Webpack comes
as a child-dependency. ESLint wants to have it as direct dependency if
it gets imported manually. But this leads to many issues in the
dependency tree. This is eventually resolvable, but we have bad
experience with version issues of vue-cli plugin sub-dependencies. Since
it is very unlikely that we will move away from using webpack (and also
then this file would not exist anymore) this will never cause any
issues.