dapp: add regular check for available version
As follow up of the recently introduced package version and version
file, this extends the application to maintain a state for the version
information.
This extends the store to know the currectly active version and the most
recently available version at the server. From that the store derives if
there is a version update available.
The active version is read from the environment variable injected durin
the packaging. The available version is regularely fetched from the
asset file at the server.

The interval to fetch the version file is done in a new class which is
called the assistent of the service worker. This might appear awkward.
This is done in preparation for future assistans to the service worker
(like sending/receiving messages). Please remind that all commits in
this area of the history are made with the knowledge of the full
update/offline feature. Thereby many things are alrady known how they
should look later.
It could be argued that fetching the version is not separated from the
actual assitants for the servie worker. On the other side the update
mechanism is tight to the service worker. The question would be where to
put this code else. In the end it is a trade-off between having a strict
responsibilty of the assistant without exceptions and a readable code
base. This mechanism needs access to the store as well. At the moment
the (planned) strucutre looks like that:
  service worker <-> assistant <-> store <-> update snackbar
It is pretty linear and easy to understand. Extracting the fetching
outside these components would lead to an unclean structure without much
gain on the other side. Therefore I think it is reasonable to let this
as an exception to the responsibility of the service worker assistant.