dapp: adopt update snackbar to use store
The update snackbar component was listining for a custom event on the
window object to recognize if an update is available. This event got
removed and the information flow that a version update is available got
reworked using the store. Therefore the component needed to adopt for
using the information from the store.

Furthermore this removes the blocking element of the component that
prevents the user to continue using the app. Since updates are optional
per default, this must not be blocking anymore.

Finally does the snackbar disconnect the SDK when the user triggers the
update if the application is in a connected state. This is necessary in
preparation to the upcoming event to make sure the SDK does not get
shoot off while being in a critical phase.