dapp: add actual update mechanism
This completes all the preparation. This makes the update snackbar
trigger the update at the service worker assistant if the user accepts
the update. The assistant notifies the service worker to do an update.
The service worker then deletes the currently cached version, unregister
itself as controller of the page and notifies the assistant to reload
the page. Note that the service worker itself has no access to the
window object nor anything else of the page. When the page gets reloaded
and no service worker exists the web-browser will automatically install
a new worker. This one will see that the cache is empty and installs the
most recent version available.