dapp: extend service worker to check cache on fail
Serving the application strictly offline without the network as fallback
requires to make sure that the cache is working. This means if the cache
gets somehow broken, the application is broken without the possibility
to recover (the user would need to clear the full cache, unregister the
service worker and reload the page). Therefore the application needs to
detect this and help/guide the user out of this situation.
Therefore a handler has been established that gets called when the
precache router of the service worker fails. Then it checks if the cache
is still okay. If that is not the case this requires to fix the cache.
It is not possible to simply load missing files (in case the cache is
partially broken), because the service could already provide a new
version. This means the only valid option is to do a full update of the
application.
Therefore the service worker notifies the user via the assistant (and
the store) that the application cache is broken an and update is
mandatory. In this case the update snackbar is using a blurry overlay to
block the application. The user should not use the application anymore
to avoid unexpected scenarios. When the user clicks on update this
behaves like a regular update.
The former case is for the scenario that the application is running
while the cache gets corrupted. The next time the user accesses a part
that is not available anymore, this will cause the handler to initiate
the mandatory update. Note that the service worker assistant, store and
update snackbar are instantiated on the initial load and are therefore
not directly affected by such a leak and functional. But in case the
application is not running while the cache gets corrupted, there is no
client to handle the mandatory update when the service worker tries to
load the application. Here the user can't be informed and asked to click
a button. The only possible way is that the service worker is doing the
update right away. Unfortunately does this still require the user to
reload the page himself. This step of the update is not working since
there is no assistant in the page its main thread to do so. But it is
expected that it is a quite usual behavior of a user to press reload if
the page does not appear. After all this is the best solution we can
come up with. (There might be the option to send desktop notifications
though.)